LGTM

+++ source/libft/check/ft_strlen.c
@@ -1,9 +1,9 @@
 
 #include "../libft.h"
 
-size_t  ft_strlen(const char *s)
+int	  ft_strlen(const char *s)
 {
-    size_t  len;
+    int  len;
 
     len = 0;
     while (s[len])
